requirements.txt
aiohttp
aiofiles
python-dotenv
websockets